+++ changelog.md
@@ -24,6 +24,7 @@ This patch adds a number of new, useful features rather than simply fixing the b
 - Faster cluster centers calculation for SMLM clusterer, DBSCAN and HDBSCAN + `lpz` is saved if applicable
 - Removed render property cache since it did not provide any significant speed improvement
 - Installers are distributed with readme.txt files (previously .rst)
+- Render: Load FOV keeps the aspect ratio of the input .txt file
 
 ## 0.10.0